rename tantivy_count_exec file
