+++ src/service/search/datafusion/plan/mod.rs
@@ -15,4 +15,4 @@
 
 pub mod deduplication;
 pub mod deduplication_exec;
-pub mod tantivy_count_exec;
+pub mod tantivy_optimize_exec;
